requirements.txt
# requirements.txt
# 游记助手项目依赖

# Streamlit 框架
streamlit>=1.28.0

# OpenAI SDK (用于 DeepSeek API)
openai>=1.0.0

# HTTP 请求
requests>=2.31.0

# 图片处理
pillow>=10.0.0

# 阿里云 OSS
oss2>=2.17.0

# 阿里云 OCR
alibabacloud-ocr-api20210707>=1.0.0
alibabacloud-tea-openapi>=0.3.7

# 阿里云 ASR (NLS)
# 注意: 阿里云 NLS SDK 可能需要额外配置，请参考官方文档

# 日期处理
python-dateutil>=2.8.2

# 数据验证
pydantic>=2.0.0
